requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
assemblyai==0.37.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
defusedxml==0.7.1
Django==5.1.6
django-cors-headers==4.7.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
dotenv==0.9.9
drf-nested-routers==0.94.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
ffmpeg==1.4
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
websockets==15.0
pdf2image
PyPDF2
pymupdf